Feat : Seperate Components - Tag, PopularLocationCards
